+++ frontend/next.config.mjs
@@ -2,6 +2,10 @@
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
+  output: "standalone",
+  experimental: {
+    appDir: true,
+  },
 };
 
 export default nextConfig;
