+++ frontend/src/hooks/useTheme.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useAtom } from "jotai";
 import { isDarkModeAtom } from "@/store/themeAtom";
 import { useEffect } from "react";
@@ -6,6 +8,9 @@ export const useTheme = () => {
   const [isDarkMode, setIsDarkMode] = useAtom(isDarkModeAtom);
 
   useEffect(() => {
+    // サーバーサイドレンダリング時は何もしない
+    if (typeof document === "undefined") return;
+
     // HTML要素のクラスを更新
     document.documentElement.classList.toggle("dark", isDarkMode);
   }, [isDarkMode]);
